新增pagehelper插件

+++ apple-orm-mybatis/src/main/java/com/appleframework/orm/mybatis/pagehelper/parser/impl/AbstractParser.java
@@ -149,8 +149,8 @@ public abstract class AbstractParser implements Parser, Constant {
         if (boundSql != null && boundSql.getParameterMappings() != null) {
             newParameterMappings.addAll(boundSql.getParameterMappings());
         }
-        newParameterMappings.add(new ParameterMapping.Builder(configuration, PAGEPARAMETER_FIRST, Integer.class).build());
-        newParameterMappings.add(new ParameterMapping.Builder(configuration, PAGEPARAMETER_SECOND, Integer.class).build());
+        newParameterMappings.add(new ParameterMapping.Builder(configuration, PAGEPARAMETER_FIRST, Long.class).build());
+        newParameterMappings.add(new ParameterMapping.Builder(configuration, PAGEPARAMETER_SECOND, Long.class).build());
         return newParameterMappings;
     }
